+++ include/rego/errors.h
@@ -8,6 +8,8 @@
 
 namespace rego
 {
+  using namespace trieste;
+
   const std::string EvalTypeError = "eval_type_error";
   const std::string EvalBuiltInError = "eval_builtin_error";
   const std::string RegoTypeError = "rego_type_error";
