+++ include/rego/rego.h
@@ -6,6 +6,7 @@
 #include "interpreter.h"
 #include "log.h"
 #include "tokens.h"
+#include "encoding.h"
 
 #include <trieste/driver.h>
 
@@ -18,6 +19,4 @@ namespace rego
   using PassCheck = std::tuple<std::string, Pass, const wf::Wellformed*>;
   std::vector<PassCheck> passes(const BuiltIns& builtins);
   Node version();
-  std::string to_json(
-    const Node& node, bool sort = false, bool rego_set = true);
 }
