+++ src/CMakeLists.txt
@@ -15,6 +15,7 @@ wf.cc
 builtins.cc
 helpers.cc
 rego_c.cc
+encoding.cc
 passes/input_data.cc
 passes/modules.cc
 passes/lists.cc
@@ -84,7 +85,9 @@ target_link_libraries(rego
 
 target_precompile_headers(rego
   PRIVATE
-    ${CMAKE_CURRENT_SOURCE_DIR}/../include/rego/rego.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/../include/rego/errors.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/../include/rego/helpers.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/../include/rego/passes.h
 )
